+++ src/bacnet/basic/object/structured_view.h
@@ -98,10 +98,24 @@ BACNET_STACK_EXPORT
 void Structured_View_Subordinate_List_Link_Array(
     BACNET_SUBORDINATE_DATA *array, size_t size);
 BACNET_STACK_EXPORT
+bool Structured_View_Subordinate_List_Element_Same(
+    BACNET_SUBORDINATE_DATA *element1, BACNET_SUBORDINATE_DATA *element2);
+BACNET_STACK_EXPORT
 BACNET_SUBORDINATE_DATA *Structured_View_Subordinate_List_Member(
     uint32_t object_instance, BACNET_ARRAY_INDEX array_index);
 BACNET_STACK_EXPORT
 unsigned int Structured_View_Subordinate_List_Count(uint32_t object_instance);
+BACNET_STACK_EXPORT
+BACNET_ARRAY_INDEX Structured_View_Subordinate_List_Element_Exist(
+    uint32_t object_instance, BACNET_SUBORDINATE_DATA *element);
+BACNET_STACK_EXPORT
+BACNET_ARRAY_INDEX Structured_View_Subordinate_List_Element_Add(
+    uint32_t object_instance, BACNET_SUBORDINATE_DATA *element);
+BACNET_STACK_EXPORT
+BACNET_ARRAY_INDEX Structured_View_Subordinate_List_Element_Remove(
+    uint32_t object_instance, BACNET_SUBORDINATE_DATA *element);
+BACNET_STACK_EXPORT
+bool Structured_View_Subordinate_List_Purge(uint32_t object_instance);
 
 BACNET_STACK_EXPORT
 BACNET_RELATIONSHIP